Clean up session queries.
